Moved CalendarUtilsTest to core module
