+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/HealersDiary.iml" filepath="$PROJECT_DIR$/HealersDiary.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/HealersDiary.iml" filepath="$PROJECT_DIR$/.idea/HealersDiary.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/app/HealersDiary.app.iml" filepath="$PROJECT_DIR$/.idea/modules/app/HealersDiary.app.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/HealersDiary.buildSrc.iml" filepath="$PROJECT_DIR$/.idea/modules/HealersDiary.buildSrc.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/core/HealersDiary.core.iml" filepath="$PROJECT_DIR$/.idea/modules/core/HealersDiary.core.iml" />
