+++ core/build.gradle.kts
@@ -3,7 +3,8 @@ import dependencies.Version
 
 plugins {
     id("com.android.library")
-    id("kotlin-android")
+    kotlin("android")
+    kotlin("kapt")
 }
 
 android {
@@ -49,6 +50,8 @@ dependencies {
     implementation(Dependencies.Firebase.analytics)
     implementation(Dependencies.timber)
     testImplementation("junit:junit:4.13.2")
+    testImplementation("com.google.truth:truth:1.1.2")
+    testImplementation("com.google.truth.extensions:truth-java8-extension:1.1.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.2")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.3.0")
     coreLibraryDesugaring("com.android.tools:desugar_jdk_libs:1.1.1")
